+++ src/lib.rs
@@ -28,6 +28,7 @@ mod macos;
 mod output;
 mod pager;
 mod pak;
+mod pkg;
 mod platform;
 mod proj;
 mod rds;